Fix: Correct JWT expirationMs type and improve imports

+++ backend/src/main/java/com/aqryuz/backend/authentication/config/JWTProperties.java
@@ -1,12 +1,13 @@
 package com.aqryuz.backend.authentication.config;
 
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+import org.springframework.validation.annotation.Validated;
+
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-import org.springframework.validation.annotation.Validated;
 
 @Component
 @ConfigurationProperties(prefix = "jwt")
@@ -15,7 +16,8 @@ import org.springframework.validation.annotation.Validated;
 @Setter
 public class JWTProperties {
 
-  @NotNull private Integer expirationMs;
+  @NotNull
+  private Long expirationMs;
 
   @NotEmpty private String secret;
 }
